+++ src/Trulioo.Client.V1/Model/RecordRule.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Trulioo.Client.V1.Model
+{
+    public class RecordRule
+    {
+        public string RuleName { get; set; }
+        public string Note { get; set; }
+    }
+}